HttpHeaders unit test refactor
